Travel expiration

+++ src/Webtaxi/MainBundle/Controller/AbstractTravelsController.php
@@ -3,17 +3,21 @@
 namespace Webtaxi\MainBundle\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
+
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpFoundation\JsonResponse;
 use Webtaxi\MainBundle\Entity\CommunicationHelper\TravelResponse;
 use Webtaxi\MainBundle\Entity\Travel;
 
 
 abstract class AbstractTravelsController extends Controller
 {
+    const STATUS_TRAVEL_ACTION_OK = 1;
+    const STATUS_TRAVEL_NOT_YOURS = -2;
+    const STATUS_TRAVEL_ALREADY_ACCEPTED = -3;
+    const STATUS_TRAVEL_IS_YOURS_CAN_NOT_ACCEPT = -4;
+    const STATUS_TRAVEL_IS_EXPIRED = -6;
+
     abstract protected function indexAction();
 
     abstract protected function getTravels($idFrom, $queryLimit);
@@ -74,17 +78,21 @@ abstract class AbstractTravelsController extends Controller
     {
         //if travel client is not current user, error:
         if ($travel->getClient() != $this->getUser()) {
-            return new Response(json_encode(array("status" => -2, "message" => "Jūs negalite trinti ne savo kelionę")));
+            return new Response(json_encode(array("status" => STATUS_TRAVEL_NOT_YOURS, "message" => "Jūs negalite trinti ne savo kelionę")));
         }
         //if traval has a driver, it could not be canceled, error:
         if ($travel->getDriver() != null) {
-            return new Response(json_encode(array("status" => -3, "message" => "Ši kelionė jau priimta, jos trinti nebegalima")));
+            return new Response(json_encode(array("status" => STATUS_TRAVEL_ALREADY_ACCEPTED, "message" => "Ši kelionė jau priimta, jos trinti nebegalima")));
+        }
+        //if travel is expired, error:
+        if ($travel->isTravelExpired()) {
+            return new Response(json_encode(array("status" => STATUS_TRAVEL_IS_EXPIRED, "message" => "Ši kelionė sukurta labai seniai. Ji nebegalioja ir jos trinti nebegalima")));
         }
 
         $em = $this->getDoctrine()->getManager();
         $em->remove($travel);
         $em->flush();
-        return new Response(json_encode(array("status" => 1, "message" => "Jūsų kelionė buvo sėkmingai ištrinta")));
+        return new Response(json_encode(array("status" => STATUS_TRAVEL_ACTION_OK, "message" => "Jūsų kelionė buvo sėkmingai ištrinta")));
 
     }
